New exercises

+++ C_05/ex05/ft_sqrt.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 int	ft_sqrt(int nb)
 {
-	long long	i;
+	long int	i;
 
-	if (nb <=0 )
+	if (nb <= 0)
 		return (0);
 	i = 1;
 	while (i * i < nb)
